Refactor sorting logic into its own class
This commit extracts sorting logic into a SortingHelper class with two methods:
  objectIntoArray(object)
  quickSort(foods)
