+++ lib/eventHandlers/foodEvents.js
@@ -8,6 +8,7 @@ $(document).ready( function () {
     Food.foodFilter();
     Food.updateFoodEvent();
     Food.updateCalorieEvent();
+    // Food.handleSortByCalories();
     $("#searchInput").on('keyup', Food.foodFilter);
     $(".form-food").on('submit', Food.addNewFood);
 });
