+++ lib/models/meal.js
@@ -2,6 +2,7 @@ const $ = require('jquery');
 const HTMLHelper = require('../helpers/htmlHelper');
 const setUrl = require('../config/apiUrl').setUrl;
 const AjaxRequests = require('../helpers/ajaxHelper');
+const SortingHelper = require('../helpers/sortingHelper');
 
 class Meal {
   static addFoodsTo(mealName) {
@@ -116,11 +117,9 @@ class Meal {
     .then((mealFoods) => {
       const foods = mealFoods.foods;
       const mealName = mealFoods.name.toLowerCase();
-      const foodsArray = new Array();
-      Object.keys(foods).map((key) => {
-        foodsArray[key] = foods[key];
-      });
-      let sortedFoods = Meal.quickSort(foodsArray);
+      
+      const foodsArray = SortingHelper.objectIntoArray(foods);
+      let sortedFoods = SortingHelper.quickSort(foodsArray);
 
       // remove the unsorted foods from the DOM
        Meal.removeMealFoodsDisplay(mealName);
@@ -142,36 +141,6 @@ class Meal {
         HTMLHelper.insertFoodsIntoMeal(mealName, foods);
       }
     })
-  }
-
-  static quickSort(foods) {
-   const foodsLength = foods.length;
-   if (foodsLength <= 1 ) {
-     // if the array only has one element, return the array
-     return foods;
-   } else if (!!foods.reduce((a, b) => { return (a.calories === b.calories) ? a : NaN })) {
-     // if every element in the array has the same value, return the array
-    return foods;
-   }
-   else {
-     const pivotIndex = Math.floor(Math.random() * foodsLength);
-     const pivotCalories = foods[pivotIndex].calories;
-     let lesserThanOrEqualToPivot = new Array();
-     let greaterThanPivot = new Array();
-     foods.forEach((food) => {
-      if (food.calories <= pivotCalories) {
-        const newIndex = lesserThanOrEqualToPivot.length;
-        lesserThanOrEqualToPivot[newIndex] = food;
-      } else {
-        const newIndex = greaterThanPivot.length
-        greaterThanPivot[newIndex] = food;
-      }
-    });
-    let sortedLesser = Meal.quickSort(lesserThanOrEqualToPivot);
-    let sortedGreater = Meal.quickSort(greaterThanPivot);
-    const sorted = sortedLesser.concat(sortedGreater);
-    return sorted;
-   }
   }
 
   static removeMealFoodsDisplay(mealName) {
